+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Support for newtype structs and byte arrays in `moonbeam-serde` form extraction.
+
 ## [0.5.0] - 2026-04-05
 
 ### Added